feat: refactor entity imports to use model package and add mappers for EntryTest and LevelScale

+++ src/main/java/fpt/org/inblue/entrytest/repository/UserCareerPreferenceRepository.java
@@ -1,10 +1,10 @@
 package fpt.org.inblue.entrytest.repository;
 
-import fpt.org.inblue.entrytest.entity.UserCareerPreference;
+import fpt.org.inblue.entrytest.model.UserCareerPreference;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserCareerPreferenceRepository extends JpaRepository<UserCareerPreference, Long> {
+public interface UserCareerPreferenceRepository extends JpaRepository<UserCareerPreference, Integer> {
     Optional<UserCareerPreference> findByUserIdAndIsActiveTrue(Integer userId);
 
     boolean existsByUserIdAndIsActiveTrue(Integer userId);
